pyproject.toml
[project]
name = "caveman-news-local"
version = "0.1.0"
description = "AI news aggregator with caveman-speak summaries"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "github-copilot-sdk>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aggregator"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
